Fix how FindBubbles handles palindromes. PopBubbles still needs to be updated.
Palindromes should be given special treatement. Palindromes are actually
a special case of bubbles formed by two strands of a node e.g the node A in
X->{A, R(A)}->R(X).

For now we don't try to pop the bubble formed by A and R(A). However,
we do add a counter so we can measure the potential impact of ignoring them.