Bug Fix. Cleaning up of temporary directories in CompressAndCorrect sometimes deleted directories we needed.
We overhauled the way we cleanup input directories. We now use the StageInfo
to determine all the output paths. We then exclude the directories we want
to keep and delete all other directories.

We mark directories as deleted by setting a modified outputpath in StageInfo
with an empty value.

* StageInfoHelper a helper class for working with StageInfo. Includes
an iterator for doing a depth first search in reverse order.

* stage_info.20131029-225613.json an actual StageInfo file used in some
  of the tests.